third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-static-relayout-expected.txt

PASS time1.offsetTop is time2.offsetTop
PASS successfullyParsed is true

TEST COMPLETE
